+++ python/pfnet/cbus.pyx
@@ -694,10 +694,6 @@ cdef class Bus:
                 g = cgen.GEN_get_next(g)
             return gens
 
-    property gens:
-        """ Same as :attr:`generators <pfnet.Bus.generators>`. """
-        def __get__(self): return self.generators
-
     property reg_generators:
         """ List of :class:`generators <pfnet.Generator>` regulating the voltage magnitude of this bus (list). """
         def __get__(self):
@@ -721,6 +717,16 @@ cdef class Bus:
                 reg_trans.append(new_Branch(br))
                 br = cbranch.BRANCH_get_reg_next(br)
             return reg_trans
+
+    property shunts:
+        """ List of :class:`shunt devices <pfnet.Shunt>` connected to this bus (list). """
+        def __get__(self):
+            shunts = []
+            cdef cshunt.Shunt* s = cbus.BUS_get_shunt(self._c_ptr)
+            while s is not NULL:
+                shunts.append(new_Shunt(s))
+                s = cshunt.SHUNT_get_next(s)
+            return shunts
 
     property reg_shunts:
         """ List of :class:`switched shunt devices <pfnet.Shunt>` regulating the voltage magnitude of this bus (list). """